Added convert.py + some customized codes for conversion
