+++ experiments/siammask_sharp_custom/config.json
@@ -53,7 +53,7 @@
     },
     "anchors": {
         "stride": 8,
-        "ratios": [0.5, 1, 2, 4, 8],
+        "ratios": [0.33, 0.5, 1, 2, 3],
         "scales": [8],
         "round_dight": 0
     }
