+++ experiments/siammask_sharp_custom/config_vot.json
@@ -6,10 +6,10 @@
         "instance_size": 255,
         "base_size": 8,
         "out_size": 127,
-        "seg_thr": 0.35,
+        "seg_thr": 0.30,
         "penalty_k": 0.04,
-        "window_influence": 0.4,
-        "lr": 1.0
+        "window_influence": 0.42,
+        "lr": 0.25
     },
     "anchors": {
         "stride": 8,
