+++ models/rpn.py
@@ -29,13 +29,17 @@ class RPN(nn.Module):
         return params
 
 
-def conv2d_dw_group(x, kernel):
-    batch, channel = kernel.shape[:2]
-    x = x.view(1, batch*channel, x.size(2), x.size(3))  # 1 * (b*c) * k * k
-    kernel = kernel.view(batch*channel, 1, kernel.size(2), kernel.size(3))  # (b*c) * 1 * H * W
-    out = F.conv2d(x, kernel, groups=batch*channel)
-    out = out.view(batch, channel, out.size(2), out.size(3))
-    return out
+class DepthwiseConv2Group(nn.Module):
+    def __init__(self):
+        super(DepthwiseConv2Group, self).__init__()
+
+    def forward(self, input, kernel):
+        batch, channel = kernel.shape[:2]
+        input = input.view(1, batch * channel, input.size(2), input.size(3))  # 1 * (b*c) * k * k
+        kernel = kernel.view(batch * channel, 1, kernel.size(2), kernel.size(3))  # (b*c) * 1 * H * W
+        feature = F.conv2d(input, kernel, groups=batch * channel)
+        feature = feature.view(batch, channel, feature.size(2), feature.size(3))
+        return feature
 
 
 class DepthCorr(nn.Module):
@@ -60,11 +64,12 @@ class DepthCorr(nn.Module):
                 nn.Conv2d(hidden, out_channels, kernel_size=1)
                 )
 
+        self.dw_conv2d_group = DepthwiseConv2Group()
+
     def forward_corr(self, kernel, input):
         kernel = self.conv_kernel(kernel)
         input = self.conv_search(input)
-        feature = conv2d_dw_group(input, kernel)
-        return feature
+        return self.dw_conv2d_group(input, kernel)
 
     def forward(self, kernel, search):
         feature = self.forward_corr(kernel, search)
