+++ tools/convert.py
@@ -1,10 +1,14 @@
-import cv2
-import glob
+# --------------------------------------------------------
+# SiamMask
+# Licensed under The MIT License
+# Written by Jiwoong Choi (jiwoong.choi@nearthlab.com)
+# --------------------------------------------------------
+import os
 import torch
 import argparse
-import numpy as np
+
 from tools.test import load_config, isfile, load_pretrain, siamese_init, siamese_track, join
-from models.siammask_sharp_custom import SiamMask
+from custom import Custom
 
 parser = argparse.ArgumentParser(description='PyTorch Tracking Demo')
 
@@ -12,7 +16,6 @@ parser.add_argument('--resume', default='', type=str, required=True,
                     metavar='PATH',help='path to latest checkpoint (default: none)')
 parser.add_argument('--config', dest='config', default='config_davis.json',
                     help='hyper-parameter of SiamMask in json format')
-parser.add_argument('--base_path', default='../../data/tennis', help='datasets')
 parser.add_argument('--cpu', action='store_true', help='cpu mode')
 args = parser.parse_args()
 
@@ -25,7 +28,7 @@ if __name__ == '__main__':
     # Setup Model
     cfg = load_config(args)
 
-    siammask = SiamMask(anchors=cfg['anchors'])
+    siammask = Custom(anchors=cfg['anchors'])
     print(type(siammask))
     if args.resume:
         assert isfile(args.resume), 'Please download {} first.'.format(args.resume)
@@ -33,43 +36,27 @@ if __name__ == '__main__':
 
     siammask.eval().to(device)
 
-    # scripted_module = torch.jit.script(siammask)
-    # print(scripted_module(torch.ones((1, 3, 127, 127)), torch.zeros((1, 3, 255, 255))))
-    # scripted_module.save('traced_siammask_model.pt')
+    outdir = './exported_models'
+    if not os.path.isdir(outdir):
+        os.mkdir(outdir)
 
-    # Parse Image file
-    img_files = sorted(glob.glob(join(args.base_path, '*.jp*')))
-    ims = [cv2.imread(imf) for imf in img_files]
+    scripted_feature_extractor = torch.jit.script(siammask.features.features)
+    scripted_feature_extractor.save(os.path.join(outdir, 'feature_extractor.pt'))
 
-    # Select ROI
-    cv2.namedWindow("SiamMask", cv2.WND_PROP_FULLSCREEN)
-    # cv2.setWindowProperty("SiamMask", cv2.WND_PROP_FULLSCREEN, cv2.WINDOW_FULLSCREEN)
-    try:
-        init_rect = cv2.selectROI('SiamMask', ims[0], False, False)
-        x, y, w, h = init_rect
-    except:
-        exit()
+    scripted_feature_downsampler = torch.jit.script(siammask.features.downsample)
+    scripted_feature_downsampler.save(os.path.join(outdir, 'feature_downsampler.pt'))
 
-    toc = 0
-    for f, im in enumerate(ims):
-        tic = cv2.getTickCount()
-        if f == 0:  # init
-            target_pos = np.array([x + w / 2, y + h / 2])
-            target_sz = np.array([w, h])
-            state = siamese_init(im, target_pos, target_sz, siammask, cfg['hp'], device=device)  # init tracker
-        elif f > 0:  # tracking
-            state = siamese_track(state, im, mask_enable=True, refine_enable=True, device=device)  # track
-            location = state['ploygon'].flatten()
-            mask = state['mask'] > state['p'].seg_thr
+    scripted_rpn_model = torch.jit.script(siammask.rpn_model)
+    scripted_rpn_model.save(os.path.join(outdir, 'rpn_model.pt'))
 
-            im[:, :, 2] = (mask > 0) * 255 + (mask == 0) * im[:, :, 2]
-            cv2.polylines(im, [np.int0(location).reshape((-1, 1, 2))], True, (0, 255, 0), 3)
-            cv2.imshow('SiamMask', im)
-            key = cv2.waitKey(1)
-            if key > 0:
-                break
+    scripted_mask_conv_kernel = torch.jit.script(siammask.mask_model.mask.conv_kernel)
+    scripted_mask_conv_kernel.save(os.path.join(outdir, 'mask_conv_kernel.pt'))
 
-        toc += cv2.getTickCount() - tic
-    toc /= cv2.getTickFrequency()
-    fps = f / toc
-    print('SiamMask Time: {:02.1f}s Speed: {:3.1f}fps (with visulization!)'.format(toc, fps))
+    scripted_mask_conv_search = torch.jit.script(siammask.mask_model.mask.conv_search)
+    scripted_mask_conv_search.save(os.path.join(outdir, 'mask_conv_search.pt'))
+
+    scripted_mask_depthwise_conv = torch.jit.script(siammask.mask_model.mask.dw_conv2d_group)
+    scripted_mask_depthwise_conv.save(os.path.join(outdir, 'mask_depthwise_conv.pt'))
+
+    scripted_refine_model = torch.jit.script(siammask.refine_model)
+    scripted_refine_model.save(os.path.join(outdir, 'refine_model.pt'))
